# docs/swagger.yaml
definitions:
...
  docker.DocEndpoint:
    properties:
      TLSConfig:
        allOf:
        - $ref: '#/definitions/docker.TLSConfiguration'
        description: TLSConfig TLS allocation
      connected:
        description: Connection status
        type: boolean
      createAt:
        type: string
      envURL:
        description: URL or IP Related to this environment（Related to this environment）Related
          to this environmentDockerRelated to this environmentURLRelated to this environmentIPRelated
          to this environment
        type: string
      id:
        description: Default parameters _id Default parameters
        type: string
      isSocket:
        description: Whether to useSocket
        type: boolean
      isTLS:
        description: Whether to useTLS
        type: boolean
      k8sConfig:
        description: K8sConfig Kubernetes allocation
        type: string
      name:
        description: Environment(Endpoint) name
        type: string
      password:
        description: Server password
        type: string
      securitySettings:
        allOf:
        - $ref: '#/definitions/docker.EndpointSecuritySettings'
        description: Environment(Endpoint) Specific security settings
      status:
        description: Environmental status（Environmental status）(1 - up, 2 - down)
        example: 1
        type: integer
      templates:
        description: All templates // All templates All templates
        items:
          $ref: '#/definitions/docker.DocTemplate'
        type: array
      type:
        description: Environment(Endpoint) 1. remotedocker api
        type: integer
      uniqueId:
        description: Server Unique Identification
        type: string
      updateAt:
        type: string
      userId:
        description: userID
        type: string
      username:
        description: Server username
        type: string
    type: object
  docker.DocRegistry:
    properties:
      AccessToken:
        description: Store temporary access token(Store temporary access token)
        type: string
      AccessTokenExpiry:
        type: integer
      authentication:
        description: Authentication Whether to enable authentication
        type: boolean
      baseURL:
        description: BaseURL in order toProGetin order to
        type: string
      createAt:
        type: string
      gitlab:
        allOf:
        - $ref: '#/definitions/docker.GitlabRegistryData'
        description: Additional fields for other warehouse types need to be stored
          in the middle(Additional fields for other warehouse types need to be stored
          in the middle，Additional fields for other warehouse types need to be stored
          in the middle3，6Additional fields for other warehouse types need to be stored
          in the middle)
      id:
        description: Default parameters _id Default parameters
        type: string
      name:
        description: Name Warehouse alias Warehouse alias
        type: string
      password:
        description: Password password
        type: string
      registryType:
        description: "Type Warehouse type\nRegistry Type. Valid values are:\n\t1 (Quay.io),
          // Not yet developed quay.io Not yet developed\n\t2 (Azure container registry),
          // Not yet developed\n\t3 (custom registry), // Self built warehouse, Self
          built warehouse\n\t4 (Gitlab registry), // Not yet developed gitlab Not
          yet developed\n\t5 (ProGet registry), // Prepare for development\n\t6 (DockerHub)
          // docker hub\n\t7 (ECR) // Not yet developed aws ECR Not yet developed"
        enum:
        - 1
        - 2
        - 3
        - 4
        - 5
        - 6
        - 7
        type: integer
      updateAt:
        type: string
      url:
        description: URL or IP Warehouse address Warehouse address Warehouse address
        type: string
      userId:
        description: Userid Reserved Usersid，Reserved Users
        type: string
      username:
        description: Username user name
        type: string
    type: object
...
  docker.DocTemplate:
    properties:
      category:
        description: Template classification
        items:
          type: string
        type: array
      config:
        allOf:
        - $ref: '#/definitions/docker.DocTemplateConfig'
        description: configuration information configuration information
      copyright:
        description: Copyright Notice
        type: string
      createAt:
        type: string
      devLanguage:
        description: development language
        items:
          type: string
        type: array
      downloads:
        description: Number of downloads
        type: integer
      features:
        description: Template Properties
        type: string
      id:
        description: Default parameters _id Default parameters
        type: string
      isDel:
        description: Delete or not
        type: boolean
      logo:
        description: softwarelogo softwarehttpsoftware
        type: string
      name:
        description: Software Name
        type: string
      nameZh:
        description: Software name in Chinese
        type: string
      notice:
        allOf:
        - $ref: '#/definitions/docker.Notice'
        description: Message notification
      poster:
        description: poster
        items:
          $ref: '#/definitions/system.SysBaseFiles'
        type: array
      preview:
        description: preview
        items:
          $ref: '#/definitions/system.SysBaseFiles'
        type: array
      relationSoft:
        description: Related software
        items:
          $ref: '#/definitions/docker.DocTemplate'
        type: array
      relationSoftIds:
        description: Related software Related software
        items:
          type: string
        type: array
      score:
        description: Average score
        type: number
      state:
        description: Software status 1. Software status 2. Software status 3. Software
          status 4. Software status
        type: integer
      supportMode:
        allOf:
        - $ref: '#/definitions/docker.Mode'
        description: Software supported modes
      sysLanguage:
        description: System Language
        items:
          type: string
        type: array
      team:
        description: the development team
        type: string
      teamIcon:
        description: the development teamIcon
        type: string
      updateAt:
        type: string
      userId:
        description: userID
        type: string
    required:
    - logo
    - name
    type: object
  docker.DocTemplateConfig:
    properties:
      createAt:
        type: string
      dockerConfig:
        additionalProperties:
          $ref: '#/definitions/docker.TemplateConfig'
        description: Dockerconfiguration information configuration informationdefault
        type: object
      dockerImage:
        description: DockerImage image
        type: string
      dockerRegistry:
        allOf:
        - $ref: '#/definitions/docker.DocRegistry'
        description: warehouseID
      fileSize:
        description: Software size
        type: number
      id:
        description: Default parameters _id Default parameters
        type: string
      isDel:
        description: Delete or not
        type: boolean
      k8SConfig:
        allOf:
        - $ref: '#/definitions/docker.KubernetesConfig'
        description: Swarmreplace content replace content
      k8SContent:
        description: SwarmFile Content
        type: string
      k8SRegistry:
        description: SwarmwarehouseID
        items:
          $ref: '#/definitions/docker.DocRegistry'
        type: array
      manual:
        allOf:
        - $ref: '#/definitions/system.SysBaseFiles'
        description: User Manual
      platform:
        description: (Reserved)Reserved
        type: integer
      swarmConfig:
        additionalProperties:
          $ref: '#/definitions/docker.TemplateConfig'
        description: Swarmreplace content replace content
        type: object
      swarmContent:
        description: SwarmFile Content
        type: string
      swarmRegistry:
        description: SwarmwarehouseID
        items:
          $ref: '#/definitions/docker.DocRegistry'
        type: array
      templateId:
        description: TemplateID
        type: string
      updateAt:
        type: string
      userId:
        description: userID
        type: string
      version:
        description: Version number
        type: string
      versionContent:
        description: Version Content
        type: string
    type: object
  docker.EndpointSecuritySettings:
    properties:
      allowBindMountsForRegularUsers:
        description: |-
          Whether non-administrator should be able to use bind mounts when creating containers
          Should non administrators be able to use binding mount when creating containers
        example: false
        type: boolean
      allowContainerCapabilitiesForRegularUsers:
        description: |-
          Whether non-administrator should be able to use container capabilities
          Should non administrators be able to use container functionality
        example: true
        type: boolean
      allowDeviceMappingForRegularUsers:
        description: |-
          Whether non-administrator should be able to use device mapping
          Should non administrators be able to use device mapping
        example: true
        type: boolean
      allowHostNamespaceForRegularUsers:
        description: |-
          Whether non-administrator should be able to use the host pid
          Should non administrators be able to use the hostpid
        example: true
        type: boolean
      allowPrivilegedModeForRegularUsers:
        description: |-
          Whether non-administrator should be able to use privileged mode when creating containers
          Should non administrators be able to use privileged mode when creating containers
        example: false
        type: boolean
      allowStackManagementForRegularUsers:
        description: |-
          Whether non-administrator should be able to manage stacks
          Should non administrators be able to manage the stack
        example: true
        type: boolean
      allowSysctlSettingForRegularUsers:
        description: |-
          Whether non-administrator should be able to use sysctl settings
          Should non administrators be able to usesysctlShould non administrators be able to use
        example: true
        type: boolean
      allowVolumeBrowserForRegularUsers:
        description: |-
          Whether non-administrator should be able to browse volumes
          Should non administrators be able to browse volumes
        example: true
        type: boolean
      enableHostManagementFeatures:
        description: |-
          Whether host management features are enabled
          Enable host management function
        example: true
        type: boolean
    type: object
  docker.Environment:
    properties:
      desc:
        type: string
      key:
        type: string
      val:
        type: string
    type: object
  docker.GitlabRegistryData:
    properties:
      instanceURL:
        type: string
      projectID:
        type: integer
      projectPath:
        type: string
    type: object
...
  docker.KubernetesConfig:
    properties:
      deployment:
        $ref: '#/definitions/v1.Deployment'
      persistentVolumeClaim:
        $ref: '#/definitions/v1.PersistentVolumeClaim'
      service:
        $ref: '#/definitions/v1.Service'
    type: object
  docker.Mode:
    properties:
      docker:
        description: Dockermode
        type: boolean
      swarm:
        description: Swarmmode
        type: boolean
    type: object
...
  docker.Notice:
    properties:
      host:
        description: ip/domain name
        type: string
      method:
        description: Request Type Request TypeGET、POST
        type: string
      query:
        description: parameter parameter version/details
        items:
          type: string
        type: array
      uri:
        description: route
        type: string
    type: object
...
  docker.Port:
    properties:
      desc:
        description: describe
        type: string
      mode:
        type: string
      protocol:
        type: string
      published:
        type: string
      target:
        type: string
    type: object
...
  docker.TLSConfiguration:
    properties:
      TLSCACertPath:
        description: TLS CAThe path to the certificate file // ca.pem
        type: string
      TLSCertPath:
        description: TLSThe path to the client certificate file // cert.pem
        type: string
      TLSKeyPath:
        description: TLSThe path to the client key file // key.pem
        type: string
      TLSType:
        description: TLSType TLStype
        example: 1
        type: integer
    type: object
  docker.TemplateConfig:
    properties:
      environment:
        items:
          $ref: '#/definitions/docker.Environment'
        type: array
      hostname:
        type: string
      ports:
        items:
          $ref: '#/definitions/docker.Port'
        type: array
      volume:
        items:
          $ref: '#/definitions/docker.VolumeMap'
        type: array
    type: object
...
  docker.VolumeMap:
    properties:
      desc:
        description: describe
        type: string
      readOnly:
        description: Read Only
        type: boolean
      source:
        description: Path to external host
        type: string
      target:
        description: Path to internal container
        type: string
      type:
        description: type bind, volume, tmpfs
        type: string
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  system.SysBaseFiles:
    properties:
      createAt:
        type: string
      id:
        description: Default parameters _id Default parameters
        type: string
      name:
        description: file name
        type: string
      originalName:
        description: original file name
        type: string
      size:
        description: file size
        type: integer
      suffix:
        description: file extension
        type: string
      updateAt:
        type: string
      url:
        description: url address
        type: string
    type: object
...
  v1.Deployment:
    properties:
      apiVersion:
        description: |-
          APIVersion defines the versioned schema of this representation of an object.
          Servers should convert recognized schemas to the latest internal value, and
          may reject unrecognized values.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources
          +optional
        type: string
      kind:
        description: |-
          Kind is a string value representing the REST resource this object represents.
          Servers may infer this from the endpoint the client submits requests to.
          Cannot be updated.
          In CamelCase.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds
          +optional
        type: string
      metadata:
        allOf:
        - $ref: '#/definitions/v1.ObjectMeta'
        description: |-
          Standard object's metadata.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata
          +optional
      spec:
        allOf:
        - $ref: '#/definitions/v1.DeploymentSpec'
        description: |-
          Specification of the desired behavior of the Deployment.
          +optional
      status:
        allOf:
        - $ref: '#/definitions/v1.DeploymentStatus'
        description: |-
          Most recently observed status of the Deployment.
          +optional
    type: object
...
  v1.DeploymentSpec:
    properties:
      minReadySeconds:
        description: |-
          Minimum number of seconds for which a newly created pod should be ready
          without any of its container crashing, for it to be considered available.
          Defaults to 0 (pod will be considered available as soon as it is ready)
          +optional
        type: integer
      paused:
        description: |-
          Indicates that the deployment is paused.
          +optional
        type: boolean
      progressDeadlineSeconds:
        description: |-
          The maximum time in seconds for a deployment to make progress before it
          is considered to be failed. The deployment controller will continue to
          process failed deployments and a condition with a ProgressDeadlineExceeded
          reason will be surfaced in the deployment status. Note that progress will
          not be estimated during the time a deployment is paused. Defaults to 600s.
        type: integer
      replicas:
        description: |-
          Number of desired pods. This is a pointer to distinguish between explicit
          zero and not specified. Defaults to 1.
          +optional
        type: integer
      revisionHistoryLimit:
        description: |-
          The number of old ReplicaSets to retain to allow rollback.
          This is a pointer to distinguish between explicit zero and not specified.
          Defaults to 10.
          +optional
        type: integer
      selector:
        allOf:
        - $ref: '#/definitions/v1.LabelSelector'
        description: |-
          Label selector for pods. Existing ReplicaSets whose pods are
          selected by this will be the ones affected by this deployment.
          It must match the pod template's labels.
      strategy:
        allOf:
        - $ref: '#/definitions/v1.DeploymentStrategy'
        description: |-
          The deployment strategy to use to replace existing pods with new ones.
          +optional
          +patchStrategy=retainKeys
      template:
        allOf:
        - $ref: '#/definitions/v1.PodTemplateSpec'
        description: Template describes the pods that will be created.
    type: object
...
  v1.ObjectMeta:
    properties:
      annotations:
        additionalProperties:
          type: string
        description: |-
          Annotations is an unstructured key value map stored with a resource that may be
          set by external tools to store and retrieve arbitrary metadata. They are not
          queryable and should be preserved when modifying objects.
          More info: http://kubernetes.io/docs/user-guide/annotations
          +optional
        type: object
      clusterName:
        description: |-
          The name of the cluster which the object belongs to.
          This is used to distinguish resources with same name and namespace in different clusters.
          This field is not set anywhere right now and apiserver is going to ignore it if set in create or update request.
          +optional
        type: string
      creationTimestamp:
        description: |-
          CreationTimestamp is a timestamp representing the server time when this object was
          created. It is not guaranteed to be set in happens-before order across separate operations.
          Clients may not set this value. It is represented in RFC3339 form and is in UTC.

          Populated by the system.
          Read-only.
          Null for lists.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata
          +optional
        type: string
      deletionGracePeriodSeconds:
        description: |-
          Number of seconds allowed for this object to gracefully terminate before
          it will be removed from the system. Only set when deletionTimestamp is also set.
          May only be shortened.
          Read-only.
          +optional
        type: integer
      deletionTimestamp:
        description: |-
          DeletionTimestamp is RFC 3339 date and time at which this resource will be deleted. This
          field is set by the server when a graceful deletion is requested by the user, and is not
          directly settable by a client. The resource is expected to be deleted (no longer visible
          from resource lists, and not reachable by name) after the time in this field, once the
          finalizers list is empty. As long as the finalizers list contains items, deletion is blocked.
          Once the deletionTimestamp is set, this value may not be unset or be set further into the
          future, although it may be shortened or the resource may be deleted prior to this time.
          For example, a user may request that a pod is deleted in 30 seconds. The Kubelet will react
          by sending a graceful termination signal to the containers in the pod. After that 30 seconds,
          the Kubelet will send a hard termination signal (SIGKILL) to the container and after cleanup,
          remove the pod from the API. In the presence of network partitions, this object may still
          exist after this timestamp, until an administrator or automated process can determine the
          resource is fully terminated.
          If not set, graceful deletion of the object has not been requested.

          Populated by the system when a graceful deletion is requested.
          Read-only.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata
          +optional
        type: string
      finalizers:
        description: |-
          Must be empty before the object is deleted from the registry. Each entry
          is an identifier for the responsible component that will remove the entry
          from the list. If the deletionTimestamp of the object is non-nil, entries
          in this list can only be removed.
          Finalizers may be processed and removed in any order.  Order is NOT enforced
          because it introduces significant risk of stuck finalizers.
          finalizers is a shared field, any actor with permission can reorder it.
          If the finalizer list is processed in order, then this can lead to a situation
          in which the component responsible for the first finalizer in the list is
          waiting for a signal (field value, external system, or other) produced by a
          component responsible for a finalizer later in the list, resulting in a deadlock.
          Without enforced ordering finalizers are free to order amongst themselves and
          are not vulnerable to ordering changes in the list.
          +optional
          +patchStrategy=merge
        items:
          type: string
        type: array
      generateName:
        description: |-
          GenerateName is an optional prefix, used by the server, to generate a unique
          name ONLY IF the Name field has not been provided.
          If this field is used, the name returned to the client will be different
          than the name passed. This value will also be combined with a unique suffix.
          The provided value has the same validation rules as the Name field,
          and may be truncated by the length of the suffix required to make the value
          unique on the server.

          If this field is specified and the generated name exists, the server will
          NOT return a 409 - instead, it will either return 201 Created or 500 with Reason
          ServerTimeout indicating a unique name could not be found in the time allotted, and the client
          should retry (optionally after the time indicated in the Retry-After header).

          Applied only if Name is not specified.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#idempotency
          +optional
        type: string
      generation:
        description: |-
          A sequence number representing a specific generation of the desired state.
          Populated by the system. Read-only.
          +optional
        type: integer
      labels:
        additionalProperties:
          type: string
        description: |-
          Map of string keys and values that can be used to organize and categorize
          (scope and select) objects. May match selectors of replication controllers
          and services.
          More info: http://kubernetes.io/docs/user-guide/labels
          +optional
        type: object
      managedFields:
        description: |-
          ManagedFields maps workflow-id and version to the set of fields
          that are managed by that workflow. This is mostly for internal
          housekeeping, and users typically shouldn't need to set or
          understand this field. A workflow can be the user's name, a
          controller's name, or the name of a specific apply path like
          "ci-cd". The set of fields is always in the version that the
          workflow used when modifying the object.

          +optional
        items:
          $ref: '#/definitions/v1.ManagedFieldsEntry'
        type: array
      name:
        description: |-
          Name must be unique within a namespace. Is required when creating resources, although
          some resources may allow a client to request the generation of an appropriate name
          automatically. Name is primarily intended for creation idempotence and configuration
          definition.
          Cannot be updated.
          More info: http://kubernetes.io/docs/user-guide/identifiers#names
          +optional
        type: string
      namespace:
        description: |-
          Namespace defines the space within which each name must be unique. An empty namespace is
          equivalent to the "default" namespace, but "default" is the canonical representation.
          Not all objects are required to be scoped to a namespace - the value of this field for
          those objects will be empty.

          Must be a DNS_LABEL.
          Cannot be updated.
          More info: http://kubernetes.io/docs/user-guide/namespaces
          +optional
        type: string
      ownerReferences:
        description: |-
          List of objects depended by this object. If ALL objects in the list have
          been deleted, this object will be garbage collected. If this object is managed by a controller,
          then an entry in this list will point to this controller, with the controller field set to true.
          There cannot be more than one managing controller.
          +optional
          +patchMergeKey=uid
          +patchStrategy=merge
        items:
          $ref: '#/definitions/v1.OwnerReference'
        type: array
      resourceVersion:
        description: |-
          An opaque value that represents the internal version of this object that can
          be used by clients to determine when objects have changed. May be used for optimistic
          concurrency, change detection, and the watch operation on a resource or set of resources.
          Clients must treat these values as opaque and passed unmodified back to the server.
          They may only be valid for a particular resource or set of resources.

          Populated by the system.
          Read-only.
          Value must be treated as opaque by clients and .
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#concurrency-control-and-consistency
          +optional
        type: string
      selfLink:
        description: |-
          SelfLink is a URL representing this object.
          Populated by the system.
          Read-only.

          DEPRECATED
          Kubernetes will stop propagating this field in 1.20 release and the field is planned
          to be removed in 1.21 release.
          +optional
        type: string
      uid:
        description: |-
          UID is the unique in time and space value for this object. It is typically generated by
          the server on successful creation of a resource and is not allowed to change on PUT
          operations.

          Populated by the system.
          Read-only.
          More info: http://kubernetes.io/docs/user-guide/identifiers#uids
          +optional
        type: string
    type: object
...
  v1.PersistentVolumeClaim:
    properties:
      apiVersion:
        description: |-
          APIVersion defines the versioned schema of this representation of an object.
          Servers should convert recognized schemas to the latest internal value, and
          may reject unrecognized values.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources
          +optional
        type: string
      kind:
        description: |-
          Kind is a string value representing the REST resource this object represents.
          Servers may infer this from the endpoint the client submits requests to.
          Cannot be updated.
          In CamelCase.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds
          +optional
        type: string
      metadata:
        allOf:
        - $ref: '#/definitions/v1.ObjectMeta'
        description: |-
          Standard object's metadata.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata
          +optional
      spec:
        allOf:
        - $ref: '#/definitions/v1.PersistentVolumeClaimSpec'
        description: |-
          Spec defines the desired characteristics of a volume requested by a pod author.
          More info: https://kubernetes.io/docs/concepts/storage/persistent-volumes#persistentvolumeclaims
          +optional
      status:
        allOf:
        - $ref: '#/definitions/v1.PersistentVolumeClaimStatus'
        description: |-
          Status represents the current information/status of a persistent volume claim.
          Read-only.
          More info: https://kubernetes.io/docs/concepts/storage/persistent-volumes#persistentvolumeclaims
          +optional
    type: object
...
  v1.Service:
    properties:
      apiVersion:
        description: |-
          APIVersion defines the versioned schema of this representation of an object.
          Servers should convert recognized schemas to the latest internal value, and
          may reject unrecognized values.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources
          +optional
        type: string
      kind:
        description: |-
          Kind is a string value representing the REST resource this object represents.
          Servers may infer this from the endpoint the client submits requests to.
          Cannot be updated.
          In CamelCase.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds
          +optional
        type: string
      metadata:
        allOf:
        - $ref: '#/definitions/v1.ObjectMeta'
        description: |-
          Standard object's metadata.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata
          +optional
      spec:
        allOf:
        - $ref: '#/definitions/v1.ServiceSpec'
        description: |-
          Spec defines the behavior of a service.
          https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status
          +optional
      status:
        allOf:
        - $ref: '#/definitions/v1.ServiceStatus'
        description: |-
          Most recently observed status of the service.
          Populated by the system.
          Read-only.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status
          +optional
    type: object
...
info:
  contact: {}
  description: This is a sample Server pets
  title: Swagger Example API
  version: 0.0.1
paths:
...
  /v2/endpoint/one/test:
    post:
      consumes:
      - application/json
      description: Testing a Single Environment
      parameters:
      - description: 'environmental information '
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/docker.DocEndpoint'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docker.DocEndpoint'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Testing a Single Environment
      tags:
      - docker
  /v2/endpoint/test:
    post:
      consumes:
      - application/json
      description: Testing the environment for local storage
      parameters:
      - description: environmental information
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/docker.DocEndpoint'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docker.DocEndpoint'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Testing the environment for local storage
      tags:
      - docker
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"